Improve test cases - Fix delete strategy over ActivityEntity.kt - Add queries to WorkoutDao.kt and WorkoutSetDao.kt - Change data type of Uid from Int to Long

+++ app/src/main/java/com/eelizarraras/workout/core/data/repository/DataBaseRepositoryImpl.kt
@@ -45,7 +45,7 @@ class DataBaseRepositoryImpl(
     }
 
     // Activity
-    override fun getActivity(uid: Int): ActivityModel? {
+    override fun getActivity(uid: Long): ActivityModel? {
         return activityDao.getActivity(uid).takeIf { it != null }?.toDomine()
     }
